chore(routes): rename /settings/wallet -> /settings/coins. resolve #1970
