--- packages/suite-desktop/pages/settings/wallet/index.tsx
@@ -1,3 +1,0 @@
-import Page from '@settings-views/wallet/Container';
-
-export default Page;
