+++ packages/suite-native/src/views/settings/coins/index.tsx
