+++ packages/suite-web/pages/settings/coins/index.tsx
@@ -1,0 +1,3 @@
+import Page from '@settings-views/coins/Container';
+
+export default Page;
