+++ packages/suite/src/views/settings/coins/Container.ts
