+++ packages/suite/src/views/settings/coins/components/Coin/index.tsx
