+++ packages/suite/src/views/settings/coins/components/CoinsGroup/index.tsx
