+++ packages/suite/src/views/settings/coins/index.tsx
